Rename 'helpers' to 'functionalHelpers' for clarity.
