+++ lib/functionalHelpers.js
