+++ lib/regularity.js
@@ -1,4 +1,4 @@
-var _             = require('./helpers');
+var _             = require('./functionalHelpers');
 var group         = require('./grouper');
 var quantify      = require('./quantifier');
